[#110] 대시보드 멤버 컴포넌트 수정
